profileSlice and userSlice integrated into survey code
